requirements.txt
setuptools~=65.6.3
einops~=0.6.1
tqdm~=4.65.0
imageio~=2.28.0
tensorboardX~=2.6
opencv-python~=4.7.0.72
matplotlib~=3.5.3
trimesh~=3.22.1
torchmetrics~=0.11.4
packaging~=23.1
ConfigArgParse~=1.5.3
Pillow~=9.3.0
pickle5~=0.0.12
kornia~=0.6.12
scipy~=1.7.2
plyfile~=0.9
wandb~=0.15.0
scikit-image~=0.19.3
pymeshlab==0.2.1
pykdtree~=1.3.7.post0
xatlas~=0.0.7
scikit-learn~=1.0.2
pytorch-lightning~=1.8.0
timm==0.6.13